browo-hrthis-backend/requirements.txt
# FastAPI Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.16

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Email Service (Brevo)
sib-api-v3-sdk==7.6.0

# File Processing
Pillow==11.0.0
python-magic==0.4.27
aiofiles==24.1.0

# Utils
pydantic==2.10.3
httpx==0.28.1
aiofiles==24.1.0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0